feat: move store folder to out of app folder
